+++ src/core/settings.rs
@@ -9,6 +9,7 @@ pub struct Settings {
     pub development_mode: bool,
     pub webdriver_url: String,
     pub root_path: String,
+    pub yolo_image_path: String,
 }
 
 lazy_static! {
@@ -22,10 +23,13 @@ lazy_static! {
             .to_string_lossy()
             .to_string();
 
+        let yolo_image_path = format!("{}/dal/local/dataset/image", root_path);
+
         Settings {
             development_mode,
             webdriver_url: webdriver_url.to_string(),
             root_path,
+            yolo_image_path,
         }
     };
 }