+++ src/domain/models/mod.rs
@@ -1,1 +1,2 @@
+pub mod ethnicity;
 pub mod yolo;