+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [unreleased]
 
+### 🚀 Features
+
+- *(models)* Created ethnicity models for url extraction
+
 ### ⚙️ Miscellaneous Tasks
 
 - *(env)* Start organize the structure of the extractor